Fix bounds for text, #125558 Patch from Martin Wehner
2004-02-11  Alexander Larsson  <alexl@redhat.com>

	* libnautilus-private/nautilus-icon-canvas-item.c (compute_text_rectangle):
	Fix bounds for text, #125558
	Patch from  Martin Wehner

+++ libnautilus-private/nautilus-icon-canvas-item.c
@@ -643,8 +643,7 @@ compute_text_rectangle (const NautilusIconCanvasItem *item,
                 text_rectangle.y0 = icon_rectangle.y0;
                 text_rectangle.y1 = text_rectangle.y0 + text_height + LABEL_OFFSET / pixels_per_unit;
 	} else {
-                text_rectangle.x0 = (icon_rectangle.x0 + icon_rectangle.x1) / 2
-			- text_width / 2;
+                text_rectangle.x0 = (icon_rectangle.x0 + icon_rectangle.x1) / 2 - (int) text_width / 2;
                 text_rectangle.y0 = icon_rectangle.y1;
                 text_rectangle.x1 = text_rectangle.x0 + text_width;
                 text_rectangle.y1 = text_rectangle.y0 + text_height + LABEL_OFFSET / pixels_per_unit;
